refactor: move game:event ipc sub to context; page uses pubsub service
